Running files and some of the directories used
This project is based on a previous work that can be found in the link: https://github.com/chrisjmccormick/hog_matlab. The code was adapt to support a 3 dimensions HOG descriptor using the color base CieLAB. Futhermore, we implemented the reidentification process by comparing all the people identified in the image with the person on the other image using color histograms.